Initial commits for projection cookbook
This cookbook provides implementations for applying projections and
selections (filters) in a variety of ways using Apache Arrow. The first
approaches to be implemented will include the use of datasets, where we
construct a ScannerBuilder and then invoke it. We will also try applying
the filter functions from the compute API along with using
`SelectColumns` from arrow::Table