task: add functional impl of naive conversion
This currently converts a list<list<...>> array to a list<...> array for
purposes of passing to existing hashing mechanisms. Nesting of data
beyond the top-level makes it complex to determine what values are on
the same row. Further, the hashing mechanisms are designed for
non-nested arrays and adding code to accommodate deep nested arrays may
be complex and error-prone. This recipe attempts to prototype an
approach that does some data manipulation independent of hashing
mechanisms.